fixed: stepping forward, stepping back, MODE_POINT, physics-bh
Stepping back was incorrect, got memcpy a bit confused. Was looking at st.size in output.c which is in bytes. MODE_POINT didn't use white. Physics-bh had its threading broken for like a few weeks now. Forgot to test. Problem was we need to allow for one variable to be negative. Forgot which. Yeah, I am going to sacrifice my external HDD and put something on it, go to a more capable computer and do some screencapturing. I figure an i7 with 4 cores and 8 threads ought to be enough to do some stuff with around 100 000. Will probably use the png dump function to get a snapshot every step.

+++ physics/physics_barnes_hut.c
@@ -82,16 +82,16 @@ static void bh_add_thread(bh_thread *root,
 		}
 	}
 	/* Split the octree and assign the threads */
-	uint8_t distrb = 8/root->assigned, remain = 8%root->assigned, oct = 0;
+	short int distrb = 8/root->assigned, remain = 8%root->assigned, oct = 0;
 	bool oct_assignments[8] = {false};
 	
-	for (uint8_t k = 1; k < root->assigned + 1; k++) {
+	for (int k = 1; k < root->assigned + 1; k++) {
 		/* Calculate the threads' share. Add the calculated remain to the first ones. */
-		uint8_t thread_conts = distrb + ((remain-- > 0) ? 1 : 0);
+		int thread_conts = distrb + ((remain-- > 0) ? 1 : 0);
 		
 		root->assign[k]->tot_octs = thread_conts;
 		
-		for (uint8_t l = 0; l < thread_conts; l++) {
+		for (int l = 0; l < thread_conts; l++) {
 			if (bh_random_assign) {
 				/* Map threads to random subtrees */
 				while (1) {
